+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Add
 
+- Add handler for correlation between log records and Open Telemetry spans (#13).
 - Add gcp.WithCallers to retrieve stack trace from error (#14).
 
 ## [0.1.1] - 2024-02-23